// web/src/api/generated/model/monthSettlementStatus.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Couplefins
 * OpenAPI spec version: 0.1.0
 */

export type MonthSettlementStatus = typeof MonthSettlementStatus[keyof typeof MonthSettlementStatus];


export const MonthSettlementStatus = {
  settled: 'settled',
  partially_settled: 'partially_settled',
  carried_forward: 'carried_forward',
} as const;